♻️ refactor: separate useGetPosts
